requirements.txt
python-telegram-bot
requests
pandas
numpy<2
python-dotenv
websocket-client
aiohttp
sqlalchemy
cryptography
matplotlib
seaborn
pyyaml
alembic
